pyproject.toml
[project]
name = "imagecaptioning"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pillow>=11.2.1",
    "requests>=2.32.4",
    "torch>=2.7.1",
    "torchvision>=0.22.1",
    "transformers>=4.52.4",
]
